% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_generic_function.R
\name{autolayer}
\alias{autolayer}
\title{Autolayer generic function}
\description{
See \code{\link[ggplot2:autolayer]{ggplot2::autolayer()}}, \code{\link[=autolayer.resilience]{autolayer.resilience()}} for details.
}
\keyword{internal}
