% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience.R
\name{autolayer.resilience}
\alias{autolayer.resilience}
\alias{autolayer.resilience_2d_Isingland}
\title{Get ggplot2 layers of resilience metrics to add to the landscape plots}
\usage{
\method{autolayer}{resilience_2d_Isingland}(
  object,
  point = TRUE,
  line = TRUE,
  split_value = TRUE,
  interval = TRUE,
  resilience_value = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{resilience} object calculated by \code{\link[=calculate_resilience]{calculate_resilience()}}}

\item{point, line, split_value, interval, resilience_value}{Show those elements on the layer? Default is \code{TRUE} for all of them.}

\item{...}{Not in use.}
}
\value{
a ggplot layer
}
\description{
Those layers can show how the resilience metrics are calculated on the landscape.
}
