% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_generic_function.R
\name{calculate_barrier}
\alias{calculate_barrier}
\title{Barrier calculation function}
\description{
See \code{\link[=calculate_barrier.Isingland]{calculate_barrier.Isingland()}} for details.
}
\keyword{internal}
