% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilience.R
\name{calculate_resilience}
\alias{calculate_resilience}
\alias{calculate_resilience.2d_Isingland}
\alias{calculate_resilience.2d_Isingland_matrix}
\alias{print.resilience_2d_Isingland}
\alias{print.resilience_2d_Isingland_matrix}
\alias{summary.resilience_2d_Isingland}
\title{Calculate the resilience values for Ising landscapes}
\usage{
calculate_resilience(l, ...)

\method{calculate_resilience}{`2d_Isingland`}(l, split_value)

\method{calculate_resilience}{`2d_Isingland_matrix`}(l, split_value, ...)

\method{print}{resilience_2d_Isingland}(x, simplify = FALSE, ...)

\method{print}{resilience_2d_Isingland_matrix}(x, ...)

\method{summary}{resilience_2d_Isingland}(object, ...)
}
\arguments{
\item{l}{An \code{Isingland} object constructed with \code{\link[=make_2d_Isingland]{make_2d_Isingland()}} or \code{\link[=make_2d_Isingland_matrix]{make_2d_Isingland_matrix()}}.}

\item{...}{Not in use.}

\item{split_value}{An integer to specify the number of active nodes used to split two resilience ranges.}

\item{x}{a result of the \emph{default} method of \code{summary()}.}

\item{simplify}{Print a simplified version of the output? Default is \code{FALSE}.}

\item{object}{an object for which a summary is desired.}
}
\description{
The resilience is calculated based on the shape of the potential landscape and the prior knowledge about the qualitatively different parts of the system. Two resilience indicators are calculated separately, and their difference is used to represent a general resilience of the system in favor of the first phase. Within each phase, the potential difference between the local maximum and the local minimum (if multiple minimums exist, use the one that is further from the other phase; and the local maximum should always be on the side to the other phase) is used to represent the resilience of this phase.
}
