% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ising_grid.R
\name{control_grid_pos_seq}
\alias{control_grid_pos_seq}
\alias{control_grid_seq}
\title{Control Functions to Specify the Varying Parameters for
an Ising Grid}
\usage{
control_grid_pos_seq(position, ...)

control_grid_seq(...)
}
\arguments{
\item{position}{The position of the single threshold or the weight
value that should vary across Ising networks. Should be a single
number for \code{"single_threshold"} or a numeric vector of
length 2 for \code{"single_wei"}.}

\item{...}{Other parameters passed to \code{\link[=seq]{seq()}} that specify the
sample grid.}
}
\description{
Functions to use within \code{\link[=make_Ising_grid]{make_Ising_grid()}}.
Use \code{control_grid_position_seq()} for types
\code{"single_threshold"} or \code{"single_wei"}; use \code{control_seq()} for
types \code{"all_thresholds"}, \code{"whole_weiadj"}, or \code{"beta"}.
}
