% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isinglandr-package.R
\docType{package}
\name{Isinglandr-package}
\alias{Isinglandr}
\alias{Isinglandr-package}
\title{Isinglandr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})

Authors:
\itemize{
  \item Gabriela Lunansky
}

}
\keyword{internal}
