% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_network_landscape.R
\name{make_2d_Isingland_matrix}
\alias{make_2d_Isingland_matrix}
\title{Make a Matrix of Landscapes for Multiple Ising Networks}
\usage{
make_2d_Isingland_matrix(Ising_grid, transform = FALSE)
}
\arguments{
\item{Ising_grid}{Parameter values generated by \code{\link[=make_Ising_grid]{make_Ising_grid()}}.}

\item{transform}{By default, this function considers the Ising network
to use \code{-1} and \code{1} for two states. Set \code{transform = TRUE} if the Ising
network uses \code{0} and \code{1} for two states.}
}
\value{

}
\description{
Make a Matrix of Landscapes for Multiple Ising Networks
}
