% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ising_grid.R
\name{make_Ising_grid}
\alias{make_Ising_grid}
\title{Make a Grid to Specify Multiple Ising Networks}
\usage{
make_Ising_grid(par1, par2 = NULL, thresholds, weiadj, beta = 1)
}
\arguments{
\item{par1, par2}{Generated from one of \code{\link[=single_threshold]{single_threshold()}},
\code{\link[=all_thresholds]{all_thresholds()}}, \code{\link[=single_wei]{single_wei()}}\verb{, [whole_weiadj()], or [beta_list()]. }par2` can be left blank.}

\item{thresholds}{Thresholds of the variables.}

\item{weiadj}{The weighted adjacency matrix.}

\item{beta}{The \eqn{\beta} value for calculating the Hamiltonian.}
}
\description{
\code{make_Ising_grid()} can specify one or two varying parameters for
Ising networks. The output of \code{make_Ising_grid()} can be used to
make landscapes of multiple networks.
}
