% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MDDNetwork}
\alias{MDDNetwork}
\alias{MDDConnectivity}
\alias{MDDThresholds}
\title{Estimation data for the Ising network of major depressive disorder}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.

An object of class \code{numeric} of length 9.
}
\source{
\url{https://figshare.com/projects/Major_depression_as_a_complex_dynamic_system_accepted_for_publication_in_PLoS_ONE_/17360}
}
\usage{
MDDConnectivity

MDDThresholds
}
\description{
Estimation data for the Ising network of major depressive disorder
}
\section{Functions}{
\itemize{
\item \code{MDDConnectivity}: The connectivity strength of the network,
represented in a weighted adjacency matrix.

\item \code{MDDThresholds}: The thresholds of the network nodes,
represented in a vector.

}}
\keyword{datasets}
