% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported_generic_function.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autolayer}
\alias{calculate_barrier}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autolayer}}}

  \item{simlandr}{\code{\link[simlandr]{calculate_barrier}}}
}}

