% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_Isingland_MDD}
\alias{shiny_Isingland_MDD}
\title{A shiny app that shows the landscape for
the Ising network of major depressive disorder}
\usage{
shiny_Isingland_MDD()
}
\description{
A shiny app that shows the landscape for
the Ising network of major depressive disorder
}
