% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_network_simulation.R
\name{simulate_2d_Isingland}
\alias{simulate_2d_Isingland}
\title{Simulate from a Landscape}
\usage{
simulate_2d_Isingland(
  l,
  mode = "single",
  initial = 0,
  length = 100,
  beta2 = l$beta
)
}
\arguments{
\item{l}{A landscape from Ising network.}

\item{mode}{One of \code{"single"}, \code{"distribution"}.}

\item{initial}{An integer or one of \code{"random"}, \code{"steady-state"}.
Float numbers will be converted to an integer automatically.}
}
\value{

}
\description{
Simulate from a Landscape
}
