% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_network_simulation.R
\name{simulate_2d_Isingland_matrix}
\alias{simulate_2d_Isingland_matrix}
\title{Title}
\usage{
simulate_2d_Isingland_matrix(
  l,
  mode = "single",
  initial = 0,
  length = 100,
  beta2 = NULL
)
}
\arguments{
\item{l}{A landscape from Ising network.}

\item{mode}{One of \code{"single"}, \code{"distribution"}.}

\item{initial}{An integer indicating the initial number
of active nodes for the simulation. Float numbers will be
converted to an integer automatically.}
}
\value{

}
\description{
Title
}
