% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_se.R
\name{calculate_stability_se}
\alias{calculate_stability_se}
\alias{print.stability_se}
\title{Calculate the standard error, confidence interval, and p-value for the stability difference between two phases of an Ising network}
\usage{
calculate_stability_se(
  data,
  split_value = 0.5 * ncol(data),
  R = 1000,
  IsingFit_options = list(plot = FALSE),
  Isingland_options = list(),
  boot.ci_options = list(type = "bca"),
  boot.pval_options = list(type = "bca"),
  ...
)

\method{print}{stability_se}(x, ...)
}
\arguments{
\item{data}{A matrix of binary data}

\item{split_value}{An integer to specify the number of active nodes used to split two stability ranges. Default is half of the number of nodes.}

\item{R}{
    The number of bootstrap replicates.  Usually this will be a single
    positive integer.  For importance resampling, some resamples may use
    one set of weights and others use a different set of weights.  In
    this case \code{R} would be a vector of integers where each
    component gives the number of resamples from each of the rows of
    weights.
  }

\item{IsingFit_options}{Parameters passed to \code{\link[IsingFit:isingfit]{IsingFit::IsingFit()}}}

\item{Isingland_options}{Parameters passed to \code{\link[=make_2d_Isingland]{make_2d_Isingland()}}}

\item{boot.ci_options}{Parameters passed to \code{\link[boot:boot.ci]{boot::boot.ci()}}. \code{type = "bca"} is set as the default because it performs best in most cases. See references for more information.}

\item{boot.pval_options}{Parameters passed to \code{\link[boot.pval:boot.pval]{boot.pval::boot.pval()}}. \code{type = "bca"} is set as the default because it performs best in most cases. See references for more information.}

\item{...}{Parameters passed to \code{\link[boot:boot]{boot::boot()}}}

\item{x}{An object of class \code{stability_se}}
}
\description{
Currently only the stability difference is calculated. The stability difference is the difference between the stability of the first phase and the stability of the second phase. The stability of a single phase is often highly zero-inflated and leads to difficulty in calculating the standard error. The stability difference is more regularly distributed and is more meaningful to perform the range estimate.
}
\references{
Puth, M.-T., Neuhäuser, M., & Ruxton, G. D. (2015). On the variety of methods for calculating confidence intervals by bootstrapping. Journal of Animal Ecology, 84(4), 892–897. https://doi.org/10.1111/1365-2656.12382
}
\seealso{
\code{\link[=compare_stability]{compare_stability()}} for comparing the stability differences of two groups.
}
