% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_se.R
\name{calculate_stability_se}
\alias{calculate_stability_se}
\alias{print.stability_se}
\title{Calculate the standard error and related range estimates for the stability metrics}
\usage{
calculate_stability_se(
  data,
  split_value = 0.5 * ncol(data),
  R = 1000,
  IsingFit_options = list(),
  Isingland_options = list(),
  boot.ci_options = list(type = "basic"),
  boot.pval_options = list(type = "basic"),
  ...
)

\method{print}{stability_se}(x, ...)
}
\arguments{
\item{data}{A matrix of binary data}

\item{split_value}{An integer to specify the number of active nodes used to split two stability ranges. Default is half of the number of nodes.}

\item{R}{
    The number of bootstrap replicates.  Usually this will be a single
    positive integer.  For importance resampling, some resamples may use
    one set of weights and others use a different set of weights.  In
    this case \code{R} would be a vector of integers where each
    component gives the number of resamples from each of the rows of
    weights.
  }

\item{IsingFit_options}{Parameters passed to \code{\link[IsingFit:isingfit]{IsingFit::IsingFit()}}}

\item{Isingland_options}{Parameters passed to \code{\link[=make_2d_Isingland]{make_2d_Isingland()}}}

\item{boot.ci_options}{Parameters passed to \code{\link[boot:boot.ci]{boot::boot.ci()}}. \code{type = "basic"} is set as the default because the stability metric tends to have zero-inflated, skewed distribution, and the basic method is the most robust to such distribution. The same reason is applied to the p-value calculation.}

\item{boot.pval_options}{Parameters passed to \code{\link[boot.pval:boot.pval]{boot.pval::boot.pval()}}}

\item{...}{Parameters passed to \code{\link[boot:boot]{boot::boot()}}}

\item{x}{An object of class \code{stability_se}}
}
\description{
Calculate the standard error and related range estimates for the stability metrics
}
