% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_se.R
\name{compare_stability}
\alias{compare_stability}
\title{Compare the stability difference between two groups}
\usage{
compare_stability(
  data1,
  data2,
  split_value = 0.5 * ncol(data),
  R = 1000,
  IsingFit_options = list(plot = FALSE),
  Isingland_options = list(),
  boot.ci_options = list(type = "bca"),
  boot.pval_options = list(type = "bca"),
  ...
)
}
\arguments{
\item{data1, data2}{Two matrices of binary data}

\item{split_value}{An integer to specify the number of active nodes used to split two stability ranges. Default is half of the number of nodes.}

\item{R}{
    The number of bootstrap replicates.  Usually this will be a single
    positive integer.  For importance resampling, some resamples may use
    one set of weights and others use a different set of weights.  In
    this case \code{R} would be a vector of integers where each
    component gives the number of resamples from each of the rows of
    weights.
  }

\item{IsingFit_options}{Parameters passed to \code{\link[IsingFit:isingfit]{IsingFit::IsingFit()}}}

\item{Isingland_options}{Parameters passed to \code{\link[=make_2d_Isingland]{make_2d_Isingland()}}}

\item{boot.ci_options}{Parameters passed to \code{\link[boot:boot.ci]{boot::boot.ci()}}. \code{type = "bca"} is set as the default because it performs best in most cases. See references for more information.}

\item{boot.pval_options}{Parameters passed to \code{\link[boot.pval:boot.pval]{boot.pval::boot.pval()}}. \code{type = "bca"} is set as the default because it performs best in most cases. See references for more information.}

\item{...}{Parameters passed to \code{\link[boot:boot]{boot::boot()}}}
}
\description{
Compare the stability difference between two groups
}
\references{
Puth, M.-T., Neuhäuser, M., & Ruxton, G. D. (2015). On the variety of methods for calculating confidence intervals by bootstrapping. Journal of Animal Ecology, 84(4), 892–897. https://doi.org/10.1111/1365-2656.12382
}
\seealso{
\code{\link[=calculate_stability_se]{calculate_stability_se()}} for getting a range estimation of the stability difference of a single group.
}
