% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_se.R
\name{compare_stability}
\alias{compare_stability}
\title{Compare the stability metrics between two Ising landscapes}
\usage{
compare_stability(
  stability_se_1,
  stability_se_2,
  boot.ci_options = list(type = "perc"),
  boot.pval_options = list()
)
}
\arguments{
\item{stability_se_1, stability_se_2}{Two objects of class \code{stability_se}}

\item{boot.ci_options}{Parameters passed to \code{\link[boot:boot.ci]{boot::boot.ci()}}}

\item{boot.pval_options}{Parameters passed to \code{\link[boot.pval:boot.pval]{boot.pval::boot.pval()}}}
}
\description{
Compare the stability metrics between two Ising landscapes
}
