% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ising_grid.R
\name{control_grid_pos_seq}
\alias{control_grid_pos_seq}
\alias{control_grid_seq}
\title{Control Functions to Specify the Varying Parameters for
an Ising Grid}
\usage{
control_grid_pos_seq(pos, seq)

control_grid_seq(seq)
}
\arguments{
\item{pos}{The position of the single threshold or the weight
value that should vary across Ising networks. Should be a single
number for \code{"single_threshold"} or a numeric vector of
length 2 for \code{"single_wei"}.}

\item{seq}{A vector that specify the values. Can be generated
with \code{\link[=seq]{seq()}}.}
}
\description{
Functions to use within \code{\link[=make_Ising_grid]{make_Ising_grid()}}.
Use \code{control_grid_pos_seq()} for types
\code{"single_threshold"} or \code{"single_wei"}; use \code{control_seq()} for
types \code{"all_thresholds"}, \code{"whole_weiadj"}, or \code{"beta"}.
}
