% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isinglandr-package.R
\docType{package}
\name{Isinglandr-package}
\alias{Isinglandr}
\alias{Isinglandr-package}
\title{Isinglandr: Landscape Construction and Simulation for Ising Networks}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A toolbox for constructing potential landscapes for Ising networks, which can be estimated by the `Isingfit` package <https://CRAN.R-project.org/package=IsingFit>. The landscapes preserve the steady-state distribution of the original model with Glauber dynamics (Glauber, 1963) <doi:10.1063/1.1703954> and can be used for visualizing and quantifying the stability of different activation states and visualizing the simulation process.
}
\author{
\strong{Maintainer}: Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})

Other contributors:
\itemize{
  \item Gabriela Lunansky [contributor]
}

}
\keyword{internal}
