% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ising_grid.R
\name{make_Ising_grid}
\alias{make_Ising_grid}
\title{Make a Grid to Specify Multiple Ising Networks}
\usage{
make_Ising_grid(
  type_par1,
  control_par1,
  type_par2 = NULL,
  control_par2 = NULL,
  thresholds,
  weiadj,
  beta = 1
)
}
\arguments{
\item{type_par1, type_par2}{One of \code{"single_threshold"},
\code{"all_thresholds"}, \code{"single_wei"}, \code{"whole_weiadj"},
or \code{"beta"}. \code{type_par2} can be left blank.}

\item{control_par1, control_par2}{See \code{\link[=control_grid_pos_seq]{control_grid_pos_seq()}}.}
}
\description{
\code{make_Ising_grid()} can specify one or two varying parameters for
Ising networks. The output of \code{make_Ising_grid()} can be used to
make landscapes of multiple networks.
}
